[FEAT] 수정페이지에서 펫 롱클릭으로 삭제 다이얼로그 띄우고 삭제 기능 구현
