+++ app/src/main/java/com/footprint/app/ui/mypage/MyPage_reviseFragment.kt
@@ -1,5 +1,8 @@
 package com.footprint.app.ui.mypage
 
+import android.app.AlertDialog
+import android.content.Context
+import android.content.DialogInterface
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.View
@@ -10,10 +13,7 @@ import com.footprint.app.R
 import com.footprint.app.databinding.FragmentMyPageReviseBinding
 
 class MyPage_reviseFragment : Fragment(R.layout.fragment_my_page_revise) {
-
-    //뷰모델과 연동
-    private lateinit var viewModel : MyPageViewModel
-
+    private lateinit var viewModel: MyPageViewModel
     private var _binding: FragmentMyPageReviseBinding? = null
     private val binding get() = _binding!!
 
@@ -21,7 +21,6 @@ class MyPage_reviseFragment : Fragment(R.layout.fragment_my_page_revise) {
         super.onViewCreated(view, savedInstanceState)
         _binding = FragmentMyPageReviseBinding.bind(view)
 
-        // ViewModel 초기화
         viewModel = ViewModelProvider(requireActivity()).get(MyPageViewModel::class.java)
 
         reUesrInfo()
@@ -32,8 +31,6 @@ class MyPage_reviseFragment : Fragment(R.layout.fragment_my_page_revise) {
 
     private fun reUesrInfo() {
         binding.save.setOnClickListener {
-
-            // 사용자가 수정한 내용을 ViewModel에 저장
             val newName = binding.mypageReviseName.text.toString()
             val newIntroduction = binding.mypageReviseIntroduction.text.toString()
             val newTown = binding.town.text.toString()
@@ -42,35 +39,59 @@ class MyPage_reviseFragment : Fragment(R.layout.fragment_my_page_revise) {
             viewModel.introduction = newIntroduction
             viewModel.town = newTown
 
-            // 마이 페이지로 이동
             findNavController().navigate(R.id.mypage)
         }
     }
 
-    private fun revisepage(){
-
-        // 수정 페이지에서 ViewModel의 데이터를 화면에 표시
+    private fun revisepage() {
         binding.mypageReviseName.setText(viewModel.name)
         binding.mypageReviseIntroduction.setText(viewModel.introduction)
         binding.town.setText(viewModel.town)
     }
 
-    private fun cance(){
+    private fun cance() {
         binding.Cance.setOnClickListener {
             findNavController().navigate(R.id.mypage)
         }
     }
 
-    private fun dog(){
-        var recyclerView = binding.dogCard
+    //리사이클러뷰에 아이템을 뛰어우고 추가 혹은 삭제 될때마다 리사이클러뷰 갱신,
+     fun dog() {
+        val recyclerView = binding.dogCard
         recyclerView.layoutManager = LinearLayoutManager(requireContext(), LinearLayoutManager.HORIZONTAL, false)
 
+        //리사이클러뷰 가로 스크롤
         val layoutManager = LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false)
         recyclerView.layoutManager = layoutManager
 
-        // 어댑터 초기화 및 dogList를 전달
-        var adapter = DogAdapter(viewModel.dogList)
+        val adapter = DogAdapter(viewModel.dogList) { position ->
+            showDeleteDialog(requireContext(), position)
+        }
         recyclerView.adapter = adapter
+    }
+
+    //삭제 다이얼로그
+    private fun showDeleteDialog(context: Context, position: Int) {
+        val builder = AlertDialog.Builder(context)
+        builder.setTitle("삭제 확인")
+        builder.setMessage("이 항목을 삭제하시겠습니까?")
+        builder.setPositiveButton("확인") { _: DialogInterface, _: Int ->
+            val adapter = binding.dogCard.adapter as DogAdapter
+            adapter.removeItem(position)
+        }
+        builder.setNegativeButton("취소") { dialog: DialogInterface, _: Int ->
+            dialog.dismiss()
+        }
+
+        //다이얼로그 취소,확인 텍스트 색상변경
+        val alertDialog = builder.create()
+        alertDialog.setOnShowListener {
+            val positiveButton = alertDialog.getButton(AlertDialog.BUTTON_POSITIVE)
+            val negativeButton = alertDialog.getButton(AlertDialog.BUTTON_NEGATIVE)
+            positiveButton.setTextColor(context.resources.getColor(R.color.black))
+            negativeButton.setTextColor(context.resources.getColor(R.color.red))
+        }
+        alertDialog.show()
     }
 
     override fun onDestroyView() {
